move closure handler to own file
