+++ src/client/callbacks.rs
@@ -137,36 +137,6 @@ pub trait ProcessHandler {
     }
 }
 
-
-/// Wrap a closure that can handle the `process` callback. This is called every time data from ports
-/// is available from JACK.
-pub struct ClosureProcessHandler<F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl> {
-    pub process_fn: F,
-}
-
-impl<F> ClosureProcessHandler<F>
-    where F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl
-{
-    pub fn new(f: F) -> ClosureProcessHandler<F> {
-        ClosureProcessHandler { process_fn: f }
-    }
-}
-
-impl<F> ProcessHandler for ClosureProcessHandler<F>
-    where F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl
-{
-    #[allow(mutable_transmutes)]
-    fn process(&mut self, c: &Client, ps: &ProcessScope) -> JackControl {
-        // Casting to mut is safe because no other callbacks will accessing the `process` field.
-        (self.process_fn)(c, ps)
-    }
-}
-
-impl<F> NotificationHandler for ClosureProcessHandler<F>
-    where F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl
-{
-}
-
 unsafe fn handler_and_ptr_from_void<'a, N, P>(ptr: *mut libc::c_void) -> &'a mut (N, P, Client)
     where N: NotificationHandler,
           P: ProcessHandler
