+++ src/client/mod.rs
@@ -2,6 +2,7 @@ mod async_client;
 mod base;
 mod callbacks;
 mod common;
+mod handler_impls;
 
 /// Contains `ClientOptions` flags used when opening a client.
 pub mod client_options;
@@ -12,6 +13,7 @@ pub mod client_status;
 pub use self::client_options::ClientOptions;
 pub use self::client_status::ClientStatus;
 pub use self::async_client::AsyncClient;
-pub use self::callbacks::{NotificationHandler, ProcessHandler, ClosureProcessHandler};
+pub use self::callbacks::{NotificationHandler, ProcessHandler};
+pub use self::handler_impls::ClosureProcessHandler;
 pub use self::base::{Client, CycleTimes, ProcessScope};
 pub use self::common::CLIENT_NAME_SIZE;
